// frontend/src/api/generated/models/CreateGreetingRequest.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * Template App API
 * API for the Template Application
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: api-support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateGreetingRequest
 */
export interface CreateGreetingRequest {
    /**
     * The greeting message content
     * @type {string}
     * @memberof CreateGreetingRequest
     */
    message: string;
    /**
     * Who the greeting is for
     * @type {string}
     * @memberof CreateGreetingRequest
     */
    recipient?: string;
}

/**
 * Check if a given object implements the CreateGreetingRequest interface.
 */
export function instanceOfCreateGreetingRequest(value: object): value is CreateGreetingRequest {
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function CreateGreetingRequestFromJSON(json: any): CreateGreetingRequest {
    return CreateGreetingRequestFromJSONTyped(json, false);
}

export function CreateGreetingRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateGreetingRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'message': json['message'],
        'recipient': json['recipient'] == null ? undefined : json['recipient'],
    };
}

export function CreateGreetingRequestToJSON(json: any): CreateGreetingRequest {
    return CreateGreetingRequestToJSONTyped(json, false);
}

export function CreateGreetingRequestToJSONTyped(value?: CreateGreetingRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'recipient': value['recipient'],
    };
}